pdfviewer: fix warnings on mac Review URL: https://codereview.chromium.org/18293014
git-svn-id: http://skia.googlecode.com/svn/trunk@9970 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ experimental/PdfViewer/SkPdfFont.h
@@ -351,8 +351,8 @@ public:
             fToUnicode = new SkPdfToUnicode(parsed, dict->ToUnicode(parsed));
         }
 
-        fFirstChar = dict->FirstChar(parsed);
-        fLastChar = dict->LastChar(parsed);
+        fFirstChar = (unsigned int)dict->FirstChar(parsed);
+        fLastChar = (unsigned int)dict->LastChar(parsed);
         fFonMatrix = dict->has_FontMatrix() ? dict->FontMatrix(parsed) : SkMatrix::I();
 
         if (dict->has_FontBBox()) {
@@ -376,7 +376,7 @@ public:
         unsigned int j = fFirstChar;
         for (unsigned int i = 0 ; i < diffs->size(); i++) {
             if ((*diffs)[i]->isInteger()) {
-                j = (*diffs)[i]->intValue();
+                j = (unsigned int)(*diffs)[i]->intValue();
             } else if ((*diffs)[i]->isName()) {
                 if (j < fFirstChar || j > fLastChar) {
                     printf("break; error 2\n");
